Modify viewer.js and app.js, create template.js
Modify viewer.js file to get a part of header(navigator) rendered with
json file by Viewer.js and create viewer.js file only to make html
templates.

A part of change from carousel.js is to get class name 'active' for
hover in style sheet due to the reason of new render of the header items and
the class name active should be assigned for one among them.